move process declaration to stylable processor
